move WithXXXX to task_opts.go
